Time: 174 ms (22.04%) | Memory: 111.8 MB (13.85%) - LeetSync

+++ 416-partition-equal-subset-sum/partition-equal-subset-sum.cpp
@@ -1,8 +1,8 @@
 class Solution {
 public:
     bool canPartition(vector<int>& nums) {
-        uint16_t sum = 0;
-        unordered_set<uint8_t> occ;
+        int sum = 0;
+        unordered_set<int> occ;
         for (size_t i = 0; i < nums.size(); i++) {
             occ.insert(nums[i]);
             sum += nums[i];
@@ -13,7 +13,7 @@ public:
         if(occ.contains(sum / 2))
             return true;
 
-        vector<vector<uint16_t>> dp(nums.size() + 1, vector<uint16_t>(sum / 2 + 1, 0));
+        vector<vector<int>> dp(nums.size() + 1, vector<int>(sum / 2 + 1, 0));
 
         for (size_t i = 1; i <= nums.size(); i++) {
             for (size_t s = 1; s <= sum / 2; s++)
